tentative modif taille fichier

+++ Packet/Packet.hpp
@@ -8,7 +8,7 @@ class Packet {
     private:
         uint8_t packetType;
         std::vector<uint8_t> data;
-        uint32_t dataSize;
+        uint64_t dataSize;
     
     public:
         Packet(std::vector<uint8_t> bytes) { fromBytes(bytes); }
@@ -19,11 +19,11 @@ class Packet {
         void fromBytes(std::vector<uint8_t> bytes);
 
         uint8_t getPacketType() { return packetType; }
-        uint32_t getDataSize() { return dataSize; }
+        uint64_t getDataSize() { return dataSize; }
         std::vector<uint8_t> getData() { return data; }
 
         void setPacketType(const uint8_t packetType) { this->packetType = packetType; }
-        void setDataSize(const uint32_t dataSize) { this->dataSize = dataSize; }
+        void setDataSize(const uint64_t dataSize) { this->dataSize = dataSize; }
         void setData(const std::vector<uint8_t> data) { this->data = data; }
         void setDataFromStr(const char* str);
